Moved view to new dir
